Reorganization: move readers for Hs.data and SQLite databases to "input_data_adapters".
Probably, most of the content from init_data.py should end up there, too (possibly in "tcga_parser.py" or something like that).
