+++ Script/BUILD
@@ -8,7 +8,7 @@ py_library(
     name = "init_data_lib",
     srcs = ["init_data.py"],
     deps = [
-        ":entrez_lookup",
+        ":hs_data_lookup",
         ":org_hs_eg_db_lookup",
     ],
 )
@@ -20,22 +20,22 @@ py_test(
 )
 
 py_library(
-    name = "entrez_lookup",
-    srcs = ["entrez_lookup.py"],
+    name = "hs_data_lookup",
+    srcs = ["input_data_adapters/hs_data_lookup.py"],
 )
 
 py_test(
-    name = "entrez_lookup_tests",
-    srcs = ["entrez_lookup_tests.py"],
+    name = "hs_data_lookup_tests",
+    srcs = ["input_data_adapters/hs_data_lookup_tests.py"],
     data = [
-        "test_data/entrez_data/first_20_Hs.data",
-        "test_data/entrez_data/first_3_short_Hs.data",
-        "test_data/entrez_data/short_Hs.data",
+        "input_data_adapters/test_data/hs_data/first_20_Hs.data",
+        "input_data_adapters/test_data/hs_data/first_3_short_Hs.data",
+        "input_data_adapters/test_data/hs_data/short_Hs.data",
     ],
-    deps = [":entrez_lookup"],
+    deps = [":hs_data_lookup"],
 )
 
 py_library(
     name = "org_hs_eg_db_lookup",
-    srcs = ["org_hs_eg_db_lookup.py"],
+    srcs = ["input_data_adapters/org_hs_eg_db_lookup.py"],
 )
